# docs/api/api.yaml
openapi: 3.0.0
info:
  title: Simple Todo API
  version: 1.0.0
paths:
  /todos:
    get:
      summary: TODOのリストを取得
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
    post:
      summary: 新しいTODOを追加
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Todo'
      responses:
        '201':
          description: TODOを追加しました
  /todos/{todoId}:
    patch:
      summary: 指定されたIDのTODOの部分的な情報を更新
      operationId: updateTodoPartial
      parameters:
        - name: todoId
          in: path
          required: true
          description: 更新するTODOのID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                subject:
                  type: string
                body:
                  type: string
      responses:
        '200':
          description: TODOの情報が正常に更新されました
        '404':
          description: TODOが見つかりませんでした
    delete:
      summary: 指定されたIDのTODOを削除
      operationId: deleteTodo
      parameters:
        - name: todoId
          in: path
          required: true
          description: 削除するTODOのID
          schema:
            type: string
      responses:
        '204':
          description: ""
        '404':
          description: TODOが見つかりませんでした
components:
  schemas:
    Todo:
      type: object
      required:
        - id
        - subject
        - body        
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        subject:
          type: string
        body:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time